+++ go.mod
@@ -1,3 +1,9 @@
-module github.com/crispy911127/edig
+module github.com/nayrsirhc/doh
 
-go 1.16
+go 1.18
+
+require (
+	github.com/inconshreveable/mousetrap v1.0.1 // indirect
+	github.com/spf13/cobra v1.6.1 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
+)
